feat(test): Move mutation tests to separate files
