+++ graphene_django_cud/tests/test_batch_delete_mutation.py
